delete with spaces bug fixed

+++ src/Socket/Socket.cpp
@@ -83,7 +83,7 @@ Socket::socket_recv()
 		if (LOG_SOCKET)
 			std::cout << "body str:" << _body_str << std::endl;
 	}
-	_request.parse_buffer(_header_str);
+	int ret = _request.parse_buffer(_header_str);
 	if (_request._request_map["Content-Type"].compare("multipart/form-data") == 0)
 	{
 		if (LOG_SOCKET)
@@ -91,7 +91,8 @@ Socket::socket_recv()
 		multipart_handler();
 		std::memset(buffer, 0, MAXLINE);
 	}
-	if (_request.get_method().compare("DELETE") == 0)
+	std::cout << "ret: " << ret << std::endl;
+	if (_request.get_method().compare("DELETE") == 0 && !ret)
 	{
 		delete_handler();
 	}
@@ -232,7 +233,3 @@ Socket::check_content_lenght_authorized()
 		}
 	}
 }
-
-/*
-	check with _server_config
-*/
